feat: /categories, /tags API개발 및 Post 기능 개발 진행중(PostService 버그 발생)

+++ src/main/java/haedal/j4/knuseminar/Domain/Category.java
@@ -13,7 +13,7 @@ public class Category {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long categoryID;  // 카테고리 ID
+    private Integer categoryID;  // 카테고리 ID
 
     @Column(length = 20, nullable = false)
     private String categoryText;  // 카테고리 내용
